fix: ensure we can still explicitly pass secrets around
Looks like I accidentally broke this. I don't really understand dagql
super well, but what I think was happening was that we would
re-evaluated the secret in the context of the new session, so passing
secrets between sessions broke.

The solution here seems to be to... pass the scope as part of the call
to secret.

Signed-off-by: Justin Chadwell <me@jedevc.com>