feat: split out go module defs into package (#6680)
* feat: split out go module defs into package

Finally! This patch splits up the codegen for go modules to produce the main
external struct defs into a separate sub-package. There are a couple of reasons
for doing this:
- Tighter control over exposed structs and their internal private methods
- Ability to import the generated structs from a sub-package

Initially this might sound simple - and prior to the addition of interfaces, it
was, however, these add a lot of confusion. Specifically, interface codegen is
added alongside the main code, and includes adding new methods on user-defined
structs as well as creating an implementation of user-defined interfaces.

So where should these new definitions belong? They actually have to go at the
same level as the definition of the struct/interface, so that 1. we can
actually attach the methods, and 2. so we can actually access the interface in
arg/return types (we can't just copy the interface into the sub-package, since
it might also refer to user-defined concrete types, which we can't copy).

But this results in a couple of complications:
- We defined a LoadFromID method for our own interfaces - we can no longer
  attach these to the client, since the client is now part of the sub-package.
  This is easy to resolve, we just make this a top-level method, it doesn't
  actually *need* to be on the client object, that was just a convention.
- We have to manually construct core types from our returns - for this, we need
  access to the `q`/`c` values on the core types. We just rename these and make
  them public, calling them `Query`/`Client` respectively. This is *annoying*
  that we have to do this at all (but there are also legitimate use cases for
  accessing these directly); as a non-urgent follow-up we should rework these
  structs so they *can't* be misused, and those fields can't be manipulated
  inappropriately.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* fix: repair the go sdk codegen

Signed-off-by: Justin Chadwell <me@jedevc.com>

* fix: re-enable foreign type checks

Signed-off-by: Justin Chadwell <me@jedevc.com>

* tests: add some go module tests for the util packages

Signed-off-by: Justin Chadwell <me@jedevc.com>

* tests: fix failing tests

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: fix vertical spacing in generated-interface code

Signed-off-by: Justin Chadwell <me@jedevc.com>

* fix: always compute the right FromID name

We were previously incorrectly computing the wrong FromID name - this
was because we weren't taking into account the current namespace. We now
take this into account, by making sure that we always get the right
namespace.

The reason this never surfaced before was because we were manually
setting the `id` field, which meant we always returned it directly, and
never actually used the query object. This is changed now though, since
we can only set `id` on local interface implementations, and not foreign
structs (since those now live in a separate package, and there's not
really a *great* case for directly exposing `id`).

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: update module_interfaces comments

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: cleanup leftover refactoring bits

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: fix docstrings

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: restore GraphQLClient

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Justin Chadwell <me@jedevc.com>