chore(sdk/python): Refactor methods that return lists to not prefetch and select id instead
Noticed a performance impact in a module when iterating over a list returned by a dependent module. The calling module was requesting a field (state), which should be returned directly by the API server, but was actually calling out (exec /runtime) on a bunch of functions that return a simple string. So these were flagged as simple fields to prefetch, even though I didn't need them, but it's the act of executing the runtime that slowed things down unnecessarily.

By loading from ID instead, it ensures that only the fields that are needed are actually requested. This is what Go does, although I think the change there was to fix a bug when lists were returned.

We could adjust the codegen to limit the prefetching selection to GraphQL fields that match Dagger module object fields, but we'd need to add a directive in the API for them when extending the schema from a module.

This is a refactor, so no additional tests were added.

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>