chore: improved error message for renderFile
This improves an absolutely hideously long error message that conceals
the main issue.

Realistically, if we get to this point, something worse has happened -
we've generated a bad dagger.gen.go. Printing the entire thing out to
stdout/stderr is not super useful for end-users (and can hide the actual
error message).

Signed-off-by: Justin Chadwell <me@jedevc.com>