fix: always set current go version on generated go.mod
Otherwise, because of `GOTOOLCHAIN=auto`, a `go.work` with a higher go
version can enforce that higher version into the new `go.mod`. Because
of this, we're then left unable to properly parse the contents of that
module (since `packages.Load` can't parse/type-check versions of go that
are higher).

We should select the go version that is explicitly packaged into the Go
SDK image, which we can do by parsing `runtime.Version()`.

Signed-off-by: Justin Chadwell <me@jedevc.com>