fix: ensure go types have MarshalJSON implementation
With the below case, the object would inherit it's implementation of
`MarshalJSON` from `Directory`:

    type Foo struct {
	    *Directory
    }

We need to define our own `MarshalJSON` so we don't end up with this
issue.

To do this, we just define another method next to our implementation of
UnmarshalJSON, using the same trick of having an intermediate concrete
struct (but we can handle module interfaces in a much simpler way, just
by storing them in a generic interface `any`, since we're only using
these for serialization, so don't need an exact type).

Signed-off-by: Justin Chadwell <me@jedevc.com>