An eclectic collection of go module codegen fixes (#6708)
* fix: codegen post-processors should error on failure

Previously, it was possible for "go mod tidy", "go work use", etc, to
error out silently. This made for some really weird to debug issues, and
also probably would have caught the issue fixed later in
4a04803cfb834c39b39ef7bac57fcf7b74c35d38.

Failing to run the post-processing commands *should* be a fatal error.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: improved error message for renderFile

This improves an absolutely hideously long error message that conceals
the main issue.

Realistically, if we get to this point, something worse has happened -
we've generated a bad dagger.gen.go. Printing the entire thing out to
stdout/stderr is not super useful for end-users (and can hide the actual
error message).

Signed-off-by: Justin Chadwell <me@jedevc.com>

* fix: always set current go version on generated go.mod

Otherwise, because of `GOTOOLCHAIN=auto`, a `go.work` with a higher go
version can enforce that higher version into the new `go.mod`. Because
of this, we're then left unable to properly parse the contents of that
module (since `packages.Load` can't parse/type-check versions of go that
are higher).

We should select the go version that is explicitly packaged into the Go
SDK image, which we can do by parsing `runtime.Version()`.

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Justin Chadwell <me@jedevc.com>