ci: always lint files even if only changes to docs/
Before this, we ignored `docs/**` for the engine lint workflow only on PRs

This leads to an inconsistency in that in PRs where you don't change
docs then you get lint errors about the docs, but if you only changes
docs then you don't see them.

Now engine lint + test workflows are split out and we ignore docs-only
changes for the tests, but always run lints.

Signed-off-by: Erik Sipsma <erik@dagger.io>