core: only allow Host.unixSocket to be used from main client (#6748)
Before this, it was technically possible for a module to make a raw
graphql query to the Host.unixSocket API and gain access to any unix
socket on the main client caller's host (i.e. where the CLI is running).
This is due to the fact that the socket attachable currently has to be
hardcoded to always refer to the main client caller (which in turn is
due to relevant grpc metadata not being forwarded by buildkit yet in
those codepaths).

Now we just enforce that this API can only ever be called from the main
client caller. This is okay for now because we don't yet support passing
unix sockets around modules. We do want to support that in the near
future, but it will require some tweaks upstream in buildkit and more
plumbing in our own code to carefully track which module has access to
a given socket based on being explicitly passed it as an arg (or
embedded in the DAG of some arg like a Container). This change just
blocks that backdoor in the meantime.

Signed-off-by: Erik Sipsma <erik@dagger.io>