engine: fix stream goroutine leak
Turns out that the goroutine used to call cancel when the grpc hijack
conn's closeCh is closed is just not necessary at all and can be
removed.

That closeCh is called when .Close is explicitly called, but in that
case there's no need to cancel our context; the context will be
cancelled anyways when the method returns.

Signed-off-by: Erik Sipsma <erik@dagger.io>