engine: fix leak of handler done log
The defer used to log "handler done" could block forever in the case
where the HandlerFunc never ends up actually being executed (i.e. some
error or disconnection happens before it even gets invoked). This is
because the chan we waited on was closed from within that HandlerFunc.

The fix here just removes the chan entirely and moves the "handler done"
log to the HandlerFunc itself, which is also much simpler anyways.

Signed-off-by: Erik Sipsma <erik@dagger.io>