fix: avoid duplicate secret accessor conversion
Fixes regression from 08cb1221f64c4496d279c849e9cf2421ef1549e2.

Essentially, it became possible to construct an input which would apply
the secret accessor translation twice. If one of the frontend inputs was
constructed from dagger, it could already have an accessor secret ID as
part of it's build graph (if `WithExec` was used alongside a secret).

However, this input could appear inside the frontend request - which
would result in double-wrapping the secret accessor to a secret that
definitely can't exist, which presents the weird error.

The easiest solution to this is to just track the inputs for dagger
calls to buildkit's Solve request, and to ensure that we avoid secret
filtering for those inputs.

Signed-off-by: Justin Chadwell <me@jedevc.com>