feat: refactor ts module internal to be more clear
Improve the TypeScript module internal to be easier to maintain
and more clear.
This refactor aims to abstract the scan logic using classes to hide
the complexity and expose a simpler interface to interact with.
By doing so, we no longer need standalone functions to do the
introspection, the class is responsible for his own introspection.

After completing the refactor, most of the scan old code with be
replaced by the new logic.

This refactor also aims to simplify some logic with the TypeScript
compiler API.

Signed-off-by: Tom Chauveau <tom@epitech.eu>