fix: propagate graphql client to child clients
This regressed in 62ced562b1bdd69f35b5112d0d2736a2d2189c7a, and meant
that child clients were invalid, and couldn't be called with `Do()`, as
well as causing `GraphQLClient()` to return an invalid nil value.

This only previously worked because the GraphQL client and the dagger
client shared the same name `client`. When this property was removed,
the propagation stopped working. Now, we make this special case
explicit.

Signed-off-by: Justin Chadwell <me@jedevc.com>