synchronously detach services on main client exit
previously service spans would be left incomplete on exit. now we'll
detach from them on shutdown, which will only stop the service if we're
the last depender on it. end result _should_ be that services are always
completed through telemetry, but I've seen maybe 2 in 50 runs still
leave it running. still troubleshooting, but without this change there
is no hope at all.

fixes #6493

Signed-off-by: Alex Suraci <alex@dagger.io>