nuclear option: wait for all spans to complete
Rather than closing the telemetry connection and hoping the timing works
out, we keep track of which traces have active spans and wait for that
count to reach 0.

A bit more complicated but not seeing a simpler solution really. Without
this we can't ensure that the client sees the very outermost spans
complete.

Signed-off-by: Alex Suraci <alex@dagger.io>