drain immediately when interrupted
otherwise we can get stuck waiting for child spans of a nested process
that got kill -9'd. not perfect but better than hanging on Ctrl+C which
is already an emergent situation where you're not likely that interested
in any remaining data if you already had a reason to interrupt.

in Cloud we'll clean up any orphaned spans based on keepalives anyway.

Signed-off-by: Alex Suraci <alex@dagger.io>