batch events that are sent from the engine
Previously we were just sending each individual update to the configured
exporters, which was very expensive and would even slow down the TUI.

When I originally tried to send it to span processors, nothing would be
sent out; turns out that was because the transform.Spans call we were
using didn't set the `Sampled` trace flag.

Now we forward engine traces and logs to all configured processors,
so their individual batching settings should be respected.

Signed-off-by: Alex Suraci <alex@dagger.io>