sdk/go: somewhat gross support for opentelemetry
had to copy-paste a lot of the telemetry code into sdk/go/. would love
to just move everything there so it can be shared between the shim, the
Go runtime, and the engine, however it is currently a huge PITA to share
code between all three, because of the way codegen works. saving that
for another day. maybe tomorrow.

Signed-off-by: Alex Suraci <alex@dagger.io>