hack: add legacy label conversion for old clients
This allows old clients like dagger v0.10.3 to still connect to and use
dagger engine v0.11.0. While this *isn't* a strict requirement, it's
nice to not break this if we don't have to, and based on a couple tests,
it seems like this was the only blocker for that.

This is needed so we can determine legacy client versions.

Signed-off-by: Justin Chadwell <me@jedevc.com>