fix: ensure client fails out when session call fails
If the main session call that was supposed to register the server fails
out, there can be no recovery! No number of retries will help, so we
should cancel the context.

Additionally, we can manipulate the error handling logic to ensure that
when the context is cancelled for this reason, that we return a
*relevant* error - we can just take the last error, this is good enough.

Signed-off-by: Justin Chadwell <me@jedevc.com>