feat: introduce minimum versioning requirements
With this change, we finally introduce minimum version requirements, in
two parts:
- Clients can enforce constraints on the minimum required server version
- Servers can enforce constraints on the minimum required client version

Doing this is a bit challenging. Both sides need to know what the other
side's version is - thankfully, we've been doing this for a *while*.

- The client can get the server version through the Buildkit client.Info
  call (where we've been attaching this data since v0.9.0).
- The server can get the client version info through old telemetry
  labels (after explicitly handling old-style telemetry labels), or
  through a newly introduced ClientMetadata field.

With this new data, we can enforce version constraints! However, in the
case of the server-side version constraints for the client, we hit a
slightly weird case - the client got stuck in a retry loop when the
Session failed to start. This isn't quite right behavior anyways (any
error from the main Session call that starts a server wouldn't cause an
immediate failure), it just appeared when we try and display version
constraints. We can easily fix this going forwards - on older clients,
we just get the failure logs, the retry will still occur, and we fail
about 5 minutes.

Signed-off-by: Justin Chadwell <me@jedevc.com>