Python: fix lock file updates (#7064)
* sdk(python): fix lock file updates

To simplify not needing to install the SDK's dependencies for codegen, the Python codegen script has been moved into an independent package with a minimal dependency.

Added a `codegen introspect` subcommand avoid the need for the Python codegen to have HTTP capabilities in order to introspect from the API directly.

This helped simplifiy the runtime module. Now, if a module has a `requirements.lock` file, it'll be re-compiled automatically (without upgrading packages). Without this, changes to the SDK's dependencies in a way that affects `import dagger`, would fail during `dagger develop`, leading to a difficult manual update of `requirements.lock`.

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>

* Move codegen tests to dedicated package

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>

* Add change log

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>

* Format

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>

* Reuse introspection

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>

* Fix lint command

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>

* Be more specific in change log

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>

---------

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>