fix: preserve windows output (#7095)
* fix: preserve windows output

Essentially, with the new TUI, we were failing to set the TUI to raw
mode. The reason for this was that we were passing a non-input buffer to
containerd/console (through os.Stdin). Unlike on *nix, on windows, the
raw mode we were trying to set is *not* interchangeable between
stdin/stdout.

Something is insanely weird about what bubbletea is doing here - it's
making *loads* of funky assumptions about how terminals can be
constructed, and honestly I'm not sure how it works. However, it seems
like we can get the right behavior as long as we preserve the
differences between input and output buffers - instead of creating one
single tty to pass around, we keep two file objects around.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: add changelog

Signed-off-by: Justin Chadwell <me@jedevc.com>

* fix: gracefully handle stdin being nil

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Justin Chadwell <me@jedevc.com>