feat: style heading in bold uppercase
As recommended by Helder, visually break sections using bold uppercase for titles.

I also removed the special template for the query subcommand: it was used to put the example after the flags. However, this created a special UX just for this sub-command. We should better have a consistent positioning of examples prior args or after on all subcommands, so on the main template

Signed-off-by: grouville <guillaume@dagger.io>