chore: update images to go 1.22
This patch updates all the go base images to versions that have go 1.22.

Note! We can't update the version used to build runc, due to an upstream
incompatability - however, because the build is separate, we can keep
building it using an older version of go.

Also note, we can't *yet* bump to actually use 1.22 in our code by
updating the go.mod. This is because the `ci` package must continue
using 1.21 until the next release (insert elaborate codegen reasons
here) - and then the `ci` package depends on the top-level package, so
that must also stay at 1.21. There *is* a potential workaround that
would allow us to split out `distconsts` into it's own go.mod, and then
we could do this, but this can be a follow-up if we really want this.

Signed-off-by: Justin Chadwell <me@jedevc.com>