fix: ensure we get correct data for prs (#7111)
Instead of messing around with a bunch of remotes, etc, we can apply
this simple little trick.

`GITHUB_REF` contains the "the fully-formed ref of the branch or tag
that triggered the workflow run" (see https://docs.github.com/en/actions/learn-github-actions/contexts#github-context).

Now, if the ref starts with "refs/pull/", it's a pull request. These
usually end in "/merge", to indicate that this is the result of merging
with the base branch. However, we can just swap this "/merge" for a
"/head" to get the actual head - then we can grab all the interesting
commit details from there.

What's really cool is that these should just all be available on the
"origin" - GitHub automagically populates them for us.

Signed-off-by: Justin Chadwell <me@jedevc.com>