fix: make TestDaggerRun pass
When we changed the format of the output to default to colored output,
this meant that we could no longer just check for `Container.from` in
the output - since we were getting control codes in there. Thankfully,
we just need to set `NO_COLOR` and we can do this check like before.

Signed-off-by: Justin Chadwell <me@jedevc.com>