refactor(sdk/elixir): move codegen from EEx to use iodata
EEx is quite hard to work on complex logic, consider return type for
example. So this changes try to convert it to `iodata` to construct
source code instead.

The advantage is we can debug the source code that construct by the
function more easily (via `IEx.pry` or `dbg`) and it's LSP friendly.

The only disadvantage is we need to add `\n` manually to construct the
right source code.

Signed-off-by: Thanabodee Charoenpiriyakij <wingyminus@gmail.com>