ci: avoid panic in sdk building
Using functions that can error in `Container.With` is a bad idea. So,
what we can do is to put the actual computation into an error group
(which has the convenient little side-effect that they all get computed
in parallel, hopefully improving build times).

This lets us remove the panic, which was crashing the binary. Because of
(I think) telemetry draining, this meant that the client job hung
forever on this kind of failure.

Signed-off-by: Justin Chadwell <me@jedevc.com>