TEMPORARY: Custom `runs-on` on manual trigger of engine-and-cli workflow
⚠️ This should be deleted / reverted before merging. It's been added for
test purposes only.

Setting `runs-on` when triggering this manually will use the same runner
type across all jobs EXCEPT testdev.

We are not allowing testdev to be changed since it requires Docker with
NVMe storage otherwise it's going to be too slow. Using a different
runner type is likely to cause failures, so leaving this one as is.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>