docs: don't use command line plaintext injection (#7232)
This is a very bad default to recommend - essentially this "de-secrets"
a value, and this could then be leaked accidentally:
- Potentially into OTEL spans
- Potentially into cache keys
- Potentially into engine logs

We should recommend using `WithSecretVariable` instead, which is much
better - then rely on the shell to do the interpolation, which avoids
all of these problems. The secret scrubber will then guarantee that we
don't secrets leaked into logs.

Signed-off-by: Justin Chadwell <me@jedevc.com>
Signed-off-by: kpenfound <kyle@dagger.io>