networking: remove no-hosts from dnsmasq conf
This setting meant that containers could not resolve names that were
resolvable in the engine (even if direct IPs were still usable).

While supporting that is incredibly obscure, the need has arisen in
integ tests for proxy support, where the test setup has:
* a proxy service
* a nested engine with dep on the proxy service
* containers running in that nested engine that automatically get proxy
  settings from their engine

In this case, being able to use the service binding alias for squid,
which is set in the engine's /etc/hosts, massively simplifies TLS
testing since the proxy's certs can all reference just that alias.

From poking around, it does not seem that this setting was load-bearing
for us and was just inherited from the original `dnsname` configuration
we picked up (and in turn `dnsname` had just set that from the beginning
with no context).

Looking around online for use cases that call for this setting in
dnsmasq, it seems to generally be obscure/complex networking setups
where entries in /etc/hosts end up pointing to IPs local to a
container/machine but they want nested clients to point to a different
IP.

My imagination is limited, but I cannot currently think of any situation
in which this would apply to engine containers; the networking settings
of the engine itself should dictate the networking settings of the
containers so any aliases found in the engine's /etc/hosts can apply to
resolving names queried by containers.
* To be clear: this does *not* result in containers' /etc/hosts having
  different values from before. It just makes all names resolvable by
  the engine also resolvable by containers, which is consistent with the
  behavior of DNS names outside of aliases.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>