chore: update dep on buildkit
The fix for leaking goroutines was fixed, so rebased our temp fork on
upstream with that.

Also, it was noticed that the fix for edge merging actually does need
the more complicated solution that includes adding jobs to all ancestor
states; there was a panic when the solver walked provenance without
that.

So this also now points to the (still pending upstream) commit with that
fix.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>