ts: fix generated code dir that's triggering unhandled diff case
The code here wasn't in and of itself wrong, but it triggered a very
subtle difference where the returned directory had subpath "/" instead
of "", which currently triggers a bug in Directory.diff.

Going to fix that bug in diff in next commit, but including this change
since once diff is fixed this new code should work in all cases and be
simpler.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>